output span attributes
