+++ exporter/debugexporter/internal/normal/traces_test.go
@@ -30,9 +30,11 @@ func TestMarshalTraces(t *testing.T) {
 				span.SetName("span-name")
 				span.SetTraceID([16]byte{0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08, 0x09, 0x0A, 0x0B, 0x0C, 0x0D, 0x0E, 0x0F, 0x10})
 				span.SetSpanID([8]byte{0x11, 0x12, 0x13, 0x14, 0x15, 0x16, 0x17, 0x18})
+				span.Attributes().PutStr("key1", "value1")
+				span.Attributes().PutStr("key2", "value2")
 				return traces
 			}(),
-			expected: `span-name 0102030405060708090a0b0c0d0e0f10 1112131415161718
+			expected: `span-name 0102030405060708090a0b0c0d0e0f10 1112131415161718 key1=value1 key2=value2
 `,
 		},
 	}
